consolidated the paging unit tests
