+++ SubSonic.Tests/DAL/ExtensionMethod/ExtensionMethodTests.cs
@@ -1,16 +1,13 @@
 ﻿using FluentAssertions;
 using NUnit.Framework;
 using SubSonic.Tests.DAL.SUT;
-using System;
-using System.Collections;
-using System.Collections.Generic;
+using System.Data.Common;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace SubSonic.Tests.DAL.ExtensionMethod
 {
     using Extensions.Test;
+    using Extensions.Test.Data.Builders;
     using SubSonic.Extensions.Test.Models;
 
     [TestFixture]
@@ -33,8 +30,61 @@ WHERE ([T1].[PersonID] = @personid_1)",
 FROM [dbo].[Renter] AS [T1]
 WHERE ([T1].[PersonID] = @personid_1)";
 
+            string
+                property_count_all = @"SELECT COUNT([T1].[ID]) [RECORDCOUNT]
+FROM [dbo].[RealEstateProperty] AS [T1]",
+                property_paged =
+@"WITH page AS
+(
+	SELECT [T1].[ID]
+	FROM [dbo].[RealEstateProperty] AS [T1]
+	ORDER BY [T1].[ID]
+	OFFSET @PageSize * (@PageNumber - 1) ROWS
+	FETCH NEXT @PageSize ROWS ONLY
+)
+SELECT [T1].[ID], [T1].[StatusID], [T1].[HasParallelPowerGeneration]
+FROM [dbo].[RealEstateProperty] AS [T1]
+	INNER JOIN page
+		ON ([page].[ID] = [T1].[ID])
+OPTION (RECOMPILE)";
+
+            Context.Database.Instance.AddCommandBehavior(property_count_all, (cmd) =>
+            {
+                using (DataTableBuilder table = new DataTableBuilder())
+                {
+                    table
+                    .AddColumn("RECORDCOUNT", typeof(int))
+                    .AddRow(RealEstateProperties.Count());
+
+                    return table.DataTable;
+                }
+            });
+            Context.Database.Instance.AddCommandBehavior(property_paged, (cmd) =>
+            {
+                int size = 0, page = 0;
+
+                foreach (DbParameter parameter in cmd.Parameters)
+                {
+                    if (parameter.ParameterName.Contains("PageSize"))
+                    {
+                        size = (int)parameter.Value;
+                    }
+                    else if (parameter.ParameterName.Contains("PageNumber"))
+                    {
+                        page = (int)parameter.Value;
+                    }
+                }
+
+                return RealEstateProperties
+                    .Skip(size * (page - 1))
+                    .Take(size)
+                    .ToDataTable();
+            });
+
+
             Context.Database.Instance.AddCommandBehavior(person_max, cmd => People.Max(x => x.ID));
             Context.Database.Instance.AddCommandBehavior(person_min, cmd => People.Min(x => x.ID));
+            
             Context.Database.Instance.AddCommandBehavior(renter_sum, cmd => Renters.Where(x => x.PersonID == cmd.Parameters["@personid_1"].GetValue<int>()).Sum(x => x.Rent));
             Context.Database.Instance.AddCommandBehavior(renter_avg, cmd => Renters.Where(x => x.PersonID == cmd.Parameters["@personid_1"].GetValue<int>()).Average(x => x.Rent));
         }
