+++ SubSonic.Tests/DAL/SubSonicContext/SubSonicContextTests.cs
@@ -340,177 +340,6 @@ WHERE ([{0}].[ID] = @id_1)";
             Context.ChangeTracking.SelectMany(x => x.Value).Count(x => x.IsNew).Should().Be(0);
         }
 
-        [Test]
-        public void ShouldBeAbleToPageData()
-        {
-            int 
-                recordCount = RealEstateProperties.Count(),
-                pageSize = 3,
-                pageCount = (int)Math.Ceiling((decimal)recordCount / pageSize);
-
-            string
-                count =
-@"SELECT COUNT([T1].[ID]) [RECORDCOUNT]
-FROM [dbo].[RealEstateProperty] AS [T1]",
-                paged =
-@"WITH page AS
-(
-	SELECT [T1].[ID]
-	FROM [dbo].[RealEstateProperty] AS [T1]
-	ORDER BY [T1].[ID]
-	OFFSET @PageSize * (@PageNumber - 1) ROWS
-	FETCH NEXT @PageSize ROWS ONLY
-)
-SELECT [T1].[ID], [T1].[StatusID], [T1].[HasParallelPowerGeneration]
-FROM [dbo].[RealEstateProperty] AS [T1]
-	INNER JOIN page
-		ON ([page].[ID] = [T1].[ID])
-OPTION (RECOMPILE)";
-
-            Context.Database.Instance.AddCommandBehavior(count, (cmd) =>
-            {
-                using (DataTableBuilder table = new DataTableBuilder())
-                {
-                    table
-                    .AddColumn("RECORDCOUNT", typeof(int))
-                    .AddRow(RealEstateProperties.Count());
-
-                    return table.DataTable;
-                }
-            });
-
-            Func<DbCommand, DataTable> command = (cmd) =>
-            {
-                int size = 0, page = 0;
-
-                foreach (DbParameter parameter in cmd.Parameters)
-                {
-                    if (parameter.ParameterName.Contains("PageSize"))
-                    {
-                        size = (int)parameter.Value;
-                    }
-                    else if (parameter.ParameterName.Contains("PageNumber"))
-                    {
-                        page = (int)parameter.Value;
-                    }
-                }
-
-                return RealEstateProperties
-                    .Skip(size * (page - 1))
-                    .Take(size)
-                    .ToDataTable();
-            };
-
-            Context.Database.Instance.AddCommandBehavior(paged, command);
-
-            IDbPageCollection<Models.RealEstateProperty> collection = Context.RealEstateProperties.ToPagedCollection(pageSize);
-
-            collection.PageSize.Should().Be(pageSize);
-
-            collection.GetRecordsForPage(1).Count().Should()
-                .BeGreaterOrEqualTo(pageSize)
-                .And
-                .BeLessThan(collection.RecordCount);
-
-            collection.PageNumber.Should().Be(1);
-
-            collection.GetRecordsForPage(2).Count().Should().Be(1);
-            collection.PageNumber.Should().Be(2);
-
-            collection.RecordCount.Should().Be(recordCount);
-            collection.PageCount.Should().Be(pageCount);
-        }
-
-        [Test]
-        public void ShouldBeAbleToEnumeratePagedData()
-        {
-            int
-                recordCount = RealEstateProperties.Count(),
-                pageSize = 3,
-                pageCount = (int)Math.Ceiling((decimal)recordCount / pageSize);
-
-            string
-                count =
-@"SELECT COUNT([T1].[ID]) [RECORDCOUNT]
-FROM [dbo].[RealEstateProperty] AS [T1]",
-                paged =
-@"WITH page AS
-(
-	SELECT [T1].[ID]
-	FROM [dbo].[RealEstateProperty] AS [T1]
-	ORDER BY [T1].[ID]
-	OFFSET @PageSize * (@PageNumber - 1) ROWS
-	FETCH NEXT @PageSize ROWS ONLY
-)
-SELECT [T1].[ID], [T1].[StatusID], [T1].[HasParallelPowerGeneration]
-FROM [dbo].[RealEstateProperty] AS [T1]
-	INNER JOIN page
-		ON ([page].[ID] = [T1].[ID])
-OPTION (RECOMPILE)";
-
-            Context.Database.Instance.AddCommandBehavior(count, (cmd) =>
-            {
-                using (DataTableBuilder table = new DataTableBuilder())
-                {
-                    table
-                    .AddColumn("RECORDCOUNT", typeof(int))
-                    .AddRow(RealEstateProperties.Count());
-
-                    return table.DataTable;
-                }
-            });
-
-            Func<DbCommand, DataTable> command = (cmd) =>
-            {
-                int size = 0, page = 0;
-
-                foreach (DbParameter parameter in cmd.Parameters)
-                {
-                    if (parameter.ParameterName.Contains("PageSize"))
-                    {
-                        size = (int)parameter.Value;
-                    }
-                    else if (parameter.ParameterName.Contains("PageNumber"))
-                    {
-                        page = (int)parameter.Value;
-                    }
-                }
-
-                return RealEstateProperties
-                    .Skip(size * (page - 1))
-                    .Take(size)
-                    .ToDataTable();
-            };
-
-            Context.Database.Instance.AddCommandBehavior(paged, command);
-
-            IDbPageCollection<Models.RealEstateProperty> collection = Context.RealEstateProperties.ToPagedCollection(pageSize);
-
-            bool enumerated = false;
-
-            foreach(IDbPageCollection<Models.RealEstateProperty> page in collection.GetPages())
-            {
-                enumerated |= true;
-
-                page.RecordCount.Should().Be(recordCount);
-
-                if (page.PageNumber == 1)
-                {
-                    page.Count().Should()
-                        .Be(pageSize)
-                        .And
-                        .BeLessThan(page.RecordCount);
-                }
-                else if (page.PageNumber == 2)
-                {
-                    page.Count().Should()
-                        .Be(1)
-                        .And
-                        .BeLessThan(page.RecordCount);
-                }
-            }
-
-            enumerated.Should().BeTrue();
-        }
+        
     }
 }
